Fixing stagePath nonsense and crashes
stagePath prevents you from opening if you don't have an unpacked rooms
folder. It also always sends you there. Now, open sends you to the last
place you opened a file if it can, to the stagePath if it can't, and to
the working dir if you have nothing unpacked.