enhance entity visual offset support
Certain entities are not placed exactly on their spawn grid square in
game.
The most prominent example of this is the creep family, which spawn
on the nearest wall. However, a number of other entities do this too.

Current list:
  - Beds
  - Satan
  - Gurdy
  - Mega Maw
  - Gate
  - Creeps

There's also a couple of entities (pokies/wall huggers and headless
horseman body) which spawn on the correct grid square but could use a
slight nudge to ensure they display in the right place. In these case,
simple image editing is not necessarily sufficient. There were a couple
of special cases that were, which have been removed by simply editing
the icon.

This change adds a `PlaceVisual` property to entities to generically
allow anything to have a numeric offset, and allow new entities that
need to use existing dynamic offsets (like `WallSnap` for creeps) to do so
without code changes. See the updates to EntitiesAfterbirthPlus for
examples.

Technically this feature can be used for more than offsets. I think it
could be cool to add some support for some other specialized entity
behaviors, like how begottens and that womb enemy connect to the nearest
wall or solid grid entity. That's the rationale for the semi-ambiguous
name.