add instant preview, fix other test methods
Instant preview from the official room editor is a much desired feature
that BR "now" supports! The reality is that it's more or less supported
it for a long time for editing anti rooms, it just needed some polish
and updates to support the ab+ launch args.

I've also enhanced and tweaked the other test options. Instead of
replacing resource files which doesn't support non-executable path
resource folders, it now creates a helper mod in your mods folder that
does all the work and disables itself when you're done.

It's worth noting that the basement test method will not work properly
if you have any mods enabled that add rooms to basement. This was the
case before and continues to be so.

The start room test mode has been broken for some time, seemingly due to
missing "essential" rooms which crashes the game on start. It now uses a
much more reliable method of loading the game's extracted special rooms
stb as the basis, which will remain up to date indefinitely.

As a last element, exporting rooms had an awful bug: if you hit cancel
on the dialog, it would save over the current stb with just that room.
I've fixed that and also enhanced the choice of starting directory for
the dialog.