add warnings for entities with out of range values
As discussed in the readme update, some entity id/variant/subtype values
are out of range.
- Types: >= 1000
- Variants: >= 4096
- Subtypes: >= 256

The editor now displays an icon over entities in the room outside this
range in any of the three values with more context on hover. It is also
logged while loading modded entities. As a bonus, the tooltip now has an
up to date x and y value, before it was just the starting x and y.