fix loading entities2 for mods without BR support
Currently BR loads the entities2 for all mods regardless of whether
autogen is on or not. This is unnecessary when it's off and the mod does
not have a BR entities xml. This change fixes that.

Also removes familiars from autogen and changes the output directory.

Also improves logging for out of bounds entities. (have had this change
locally for a while, it's minor but it helps for debugging and finding
these rooms)