add save and test hooks, BR mod callbacks, enhance test methods
Hooks are a list of scripts that get called when a BR action happens.
This implements hooks for saving and starting a room test.

To go with the test hooks, the BR helper mod is now formalized and has a
lua component that has callbacks. This is the first half of
compatibility with stage api.

Improves the basement replacement test method; it now can replace
whatever floor's files are relevant. It and starting room will now also
change stage as appropriate when the game starts.

Stage Replacement tests also now force the floor to be XL, turns off
Curse of the Lost, and gives the player The Mind to make reset scumming
easier.

Instapreview now supports steam urls

Add blue womb and fix cathedral and sheol for instapreview; the ids
weren't updated after rebirth and blue womb took their old id