add bulk replace entities dialog
Adds the ability to replace all entities of a given type in all rooms in a
file in a single action. Edit > Bulk Replace Entities

The resulting dialog will pull the replaced entity from the room
selection, and the replacement from the palette selection. It displays a
preview icon of the entities

Feature creep: this also adds a new icon indicating custom entities with
missing entities2 entries, a red version of the similar yellow warning.
This was deemed necessary as these entities crash the game on start and
it was only reported in the logs until now

Converts Entity non-rendering data into a standalone class so it can be
used by other things that don't care about editor logic

Fixes a bug in saving that prevented save hooks from opening the output
stb; it was keeping the file open past the necessary save point

Adds a niche feature list, so we can document niche BR things that
aren't obvious without looking at the code or tribal knowledge.