enhance autogen mod icon support
Previously, autogen used the first frame of the first layer of the
default animation of an entity as its icon. It was suggested that it
should actually use every layer's first frame in that animation

This recreates the transformational properties that an anm2 frame can
have in more depth in order to faithfully reassemble the first frame of
animation. Note this will still not work for entities that use overlays,
or for entities with only invisible frames/layers

Also adds a helper setting, FixIconFormat, which allows users to fix
annoying libpng warnings clogging up the logs