fix fallback logic for finding the mod path
Instead of falling back properly when the install folder is invalid,
finding the mod path gave up pretty early. This fixes that.

Fix a bug that happened due to the ModTemp entities folder not
being committed to git. This just generates it when autogen is on

Fix a bug caused by a stash in the previous entity info conversion

Add an error message for when a mod entity has a duplicate id, variant,
and subtype

Add baseHP to autogenerated entities