add option to disable test dialog
Now the modal popup when testing can be disabled. If it is, 5 minutes
after starting a test without beginning another one the helper mod will
be disabled. It will also be unconditionally disabled upon closing
normally

The helper mod now displays on screen text to clearly indicate that it's
enabled

Fix some bugs in room duplication, they're now inserted in the right
place and with better ids

Fix a crash when selecting an invalid entity in a set

Add a room prefix to the log indicating invalid entities in a room file

Fix crash when opening file that does not exist