add rebirth stb compatibility
This isn't full rebirth compatibility, as it still shows all the
afterbirth ents and doesn't account for adjusted entity ids.
This would be a much bigger undertaking, (particularly requiring
and overhaul in how the data xmls are handled) and may be done later.

add fallback shape 0 code for afterbirth stbs with null rooms