antibirth stb compatibility
Antibirth uses the afterbirth format in most places, but occasionally it uses
a modified version of it with 9 extra bytes in the room header.

In all the files I looked at, this was unused, but it may be more clear
in rep. I don't have the anti entities xml or anything so that will
rely on someone else making it.