Entity grouping overhaul draft
This replaces the Kind and Group systems in Basement Renovator as well as all associated hardcoding with a new unified Group system, allowing easily defining new tabs and the ordering of said tabs.

Currently I've only done minor work on the xml itself, but this should function as a proof of concept.